+++ src/lib/components/Header.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+	import { page } from '$app/stores';
+
+	// Components
+	import { Logo } from '$components';
+</script>
+
+<header>
+	<Logo />
+	Amsterdam
+</header>
