--- src/lib/components/Navigation.svelte
@@ -1,10 +1,0 @@
-<script lang="ts">
-	import { Logo } from '$components';
-
-	export let title: string = 'Navigation';
-</script>
-
-<nav>
-	<Logo />
-	{title}
-</nav>
