+++ src/lib/components/index.ts
@@ -1,3 +1,4 @@
-export { default as Navigation } from './Navigation.svelte';
+export { default as Header } from './Header.svelte';
 export { default as Logo } from './Logo.svelte';
 export { default as LocaleSwitcher } from './LocaleSwitcher.svelte';
+export { default as Footer } from './Footer.svelte';
